fixed colors and added delete confirmation

+++ src/components/list-item/list-item.js
@@ -7,10 +7,11 @@ export default {
         return {
             editMode: false,
             loading: false,
+            errors: [],
             newRecord: {
                 title: this.itemData.title,
                 bodyText: this.itemData.bodyText,
-                authorId: this.itemData.authorId,
+                authorId: 1,
                 lirixId: this.itemData.lirixId
             }
         }
@@ -22,25 +23,67 @@ export default {
             types.updateData
         ]),
         deleteItemfromDatabase(postId) {
-            this.loading = true;
-            this.deleteData(postId)
-                .then(() => {
-                    this.loading = false;
-                });
+
+            if (confirm('Are you sure?')) {
+                this.loading = true;
+
+                this.deleteData(postId)
+                    .then(() => {
+                        this.loading = false;
+                    });
+            }
+
         },
         updateItemInDatabase() {
-            this.loading = true;
-            this.updateData(this.newRecord)
-                .then(() => {
-                    this.loading = false;
-                    this.disableEditing();
-                });
+
+
+
+            let fieldData = this.newRecord;
+
+            if (fieldData.title && fieldData.bodyText && fieldData.authorId) {
+                this.loading = true;
+                this.updateData(this.newRecord)
+                    .then(() => {
+                        this.loading = false;
+                        this.disableEditing();
+                    });
+            }
+
+            this.errors = [];
+
+            if (!fieldData.title) {
+                this.errors.push('Add a title');
+            }
+            if (!fieldData.authorId) {
+                this.errors.push('Select an author ');
+            }
+            if (!fieldData.bodyText) {
+                this.errors.push('Add some content');
+            }
         },
         enableEditing() {
             this.editMode = true
         },
         disableEditing() {
             this.editMode = false
+        }
+    },
+    computed: {
+        addedOnDate() {
+            let now = new Date().getDay();
+            let date = new Date(this.itemData.timestamp);
+
+            let diffTime = Math.abs(now - date);
+            let diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+
+
+
+            if (diffDays > 1) {
+                return ('at ' + date.getHours() + ':' + date.getMinutes());
+            } else {
+                return ('on ' + date.getDate() + "-" + (date.getMonth() + 1) + "-" + date.getFullYear());
+            }
+
         }
     }
 }
